Ajuste no service worker
