+++ index.html
@@ -63,6 +63,10 @@
         </div>
       </div>
     </main> 
+    
+    <!-- import the webpage's javascript file -->
+    <script src="script.js" defer></script>
+
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
   </body>
 </html>
