+++ site.webmanifest
@@ -5,12 +5,12 @@
     "short_name": "musicaIPRV",
     "icons": [
         {
-            "src": "/docs/favicon/android-chrome-192x192.png",
+            "src": "docs/favicon/android-chrome-192x192.png",
             "sizes": "192x192",
             "type": "image/png"
         },
         {
-            "src": "/docs/favicon/android-chrome-512x512.png",
+            "src": "docs/favicon/android-chrome-512x512.png",
             "sizes": "512x512",
             "type": "image/png"
         }
